Switch to port 8080
Port 8080 should work anywhere and everywhere and will now be used to serve up everything from the server (flask).
A future update could use 8080 as a default port number but allow for other options using argparse.

Port 8080 has long been used as an alternate port for HTTP and works without the need of any system priveleges (port 80 sometimes does). Therefore, it should work in any environment.
As an added bonus, some environments even redirect traffic from port 80 to port 8080 meaning that the webpage can be visited without using any port numbers.
In  environments where this is not aleady the case, it is easy to set up (but not at all a requirement).

All references to port 5000 (there was one inside a comment) have been updated and/or removed.