Fixed X/Y coordinate swapping
The event logging has been drastically improved as well.
Both the client and the server now have an event log with the coordinates sent, and the coordinates received.
For clarity, both the server and the client print out the information in the smae format with CLEAR labels explaining what is going on.

This should make solving similar issues much easier in the future.