Functionalized move setting
The proccess of checking a moves validity, applying it, and applying any KO has now been turned into a single function.
Besides simplifying the code and providing a standardized error method, this change also reduces the size of the codebase.

Additionally, moves can be applied out of turn which should make it possible to apply handicap moves or to allow one game to be played across multiple devices.